backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
sqlalchemy
aiosqlite
google-generativeai
numpy
requests
scikit-learn
pandas
joblib
python-multipart

